Download PADRINO pdb objects

Download PADRINO from Github.

pdb_download(save = TRUE, destination = NULL)

pdb_save(pdb, destination = NULL)

pdb_load(path)

Arguments

save

Write the PDB object to a folder of text files?

destination

Where to write the pdb object to.

pdb

A pdb object.

path

The directory where the PADRINO tables are stored

Value

pdb_download and pdb_load return pdb objects. pdb_save returns a pdb object invisibly.

Details

This does not currently support versioning because there is only one version. destination should be a folder name. When save = TRUE, a set of 12 text files will be saved in the destination folder. The files are tab-delimited.