Generate IPMs from Padrino objects

This function generates complete IPMs from objects created with pdb_make_proto_ipm.

pdb_make_ipm(proto_ipm_list, addl_args = list())

Arguments

proto_ipm_list
Output from pdb_make_proto_ipm.

addl_args
A named list of additional arguments to pass to make_ipm.

Value

A list of IPMs.

Details

The format of addl_args should be a nested list. The names of the outermost level should correspond to the ipm_id that the arguments apply to. Each entry of the outermost level should itself then be a named list where the names correspond to arguments to make_ipm, and the values are the values for each argument. See examples.

Examples

if (FALSE) {

data("pdb_ex")

proto <- pdb_make_proto_ipm(pdb_ex, ipm_id = "aaa341", det_stoch = "det")

ipm   <- pdb_make_ipm(proto)

proto <- pdb_make_proto_ipm(pdb_ex,
                            ipm_id     = "aaaa55",
                            det_stoch  = "stoch",
                            kern_param = "kern")

args  <-list(

 # The names in the outermost list should be ipm_id's

 aaaa55 = list(

   # The names in the inner list should be arguments to make_ipm()

   report_progress = TRUE,
   iterate         = TRUE,
   iterations      = 100,
   kernel_seq      = sample(2004:2014, 100, replace = TRUE)
 )
)

ipm   <- pdb_make_ipm(proto, addl_args = args)
}