Generate proto_ipms from Padrino objects

This function generates proto_ipm objects from Padrino Database tables.

pdb_make_proto_ipm(
  pdb,
  ipm_id = NULL,
  det_stoch = "det",
  kern_param = "kern",
  stop_on_failure = TRUE
)

Arguments

pdb
A pdb object.

ipm_id
Optionally, one or more ipm_id's to build. If empty, all models contained in the pdb object will be processed into proto_ipm's.

det_stoch
A vector containing either "det" or "stoch". This determines whether we want to construct a deterministic or stochastic model. Default is "det". See details

kern_param
If det_stoch = "stoch", then whether or not to construct a kernel resampled model, or a parameter resampled model. See details.

stop_on_failure
A logical. If TRUE, when building many models, will halt the process with an error if any one of them is unable to build with the requested parameters. Otherwise, it will throw a warning indicating which models cannot be built.

Value

A list containing one or more proto_ipms. Names of the list will correspond to ipm_ids.

Details

proto_ipm objects contain all of the information needed to implement an IPM, but stop short of actually generating kernels. These are intermediate building blocks that can be modified before creating a full IPM so that things like perturbation analysis are a bit more straightforward.

When requesting many models, the det_stoch and kern_param parameters can also be vectors. These are matched with ipm_id by position. If the lengths of det_stoch and kern_param do not match the length ipm_id, they will be recycled until they do.

For stochastic models, there is sometimes the option of building either a kernel-resampled or a parameter resampled model. A kernel resampled model uses some point estimate for time and/or space varying parameters to generate kernels for each year/site/grouping factor. Parameter resampled models sample parameters from distributions. Padrino stores this information for some models when it is available in the literature, and tries to fail informatively when these distributions aren't available in the database.

See also